requirements.txt
kafka-python==2.2.15
psycopg2==2.9.10
